requirements.txt
anyio==3.7.0
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
config==0.5.1
cryptography==41.0.1
Django==4.2.2
django-request==1.6.2
exceptiongroup==1.1.1
fastapi==0.98.0
fasttext-wheel
h11==0.14.0
idna==3.4
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
pycparser==2.21
pydantic==1.10.9
pytesseract
stanza
python-dateutil==2.8.2
python-multipart==0.0.6
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
typing-extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
Wand==0.6.11
Levenshtein
scipy
transformers